Trigger pyre extension when pyre config is anywhere in workspace
Summary:
We were not attempting to start the pyre vscode extension if the pyre configuration did not live at the root of the workspace, which was confusing/hard to debug because there were no output logs, etc.

Let's make this flexible and start as long as there is a config anywhere.

Docs on activation triggers for vscode extensions: https://code.visualstudio.com/api/references/activation-events#workspaceContains

Reviewed By: stroxler

Differential Revision: D35825671

fbshipit-source-id: 6da947dc92da4d5841f4a2251324594e8079e9a5